pyproject.toml
[project]
name = "verity"
version = "0.1.0"
description = "Verity is the code name for an unnamed project about budgeting and financial planning"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "pylimbo>=0.0.19",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
